Open up TransportationResult and CommunicatorResponse types
