+++ Sources/Endpoints/Transporter/Transporter.swift
@@ -1,6 +1,6 @@
 //
 //  Transporter.swift
-//  Endpoints iOS
+//  Endpoints
 //
 //  Created by Simon Jarbrant on 2019-05-25.
 //
@@ -9,9 +9,4 @@ import Foundation
 
 public protocol Transporter {
     func send(_ request: URLRequest, completionHandler: @escaping (Result<TransportationResult, Error>) -> Void) -> Cancellable
-}
-
-public struct TransportationResult {
-    let response: HTTPURLResponse
-    let data: Data
 }
